Refactored BlockExplorerLink variations to be included on base component
