+++ src/apps/gp-v1/components/common/BlockExplorerLink/index.tsx
@@ -1,14 +1,3 @@
-import React from 'react'
+import { BlockExplorerLinkWalletConnection } from 'components/common/BlockExplorerLink'
 
-import { BlockExplorerLink, Props } from 'components/common/BlockExplorerLink'
-
-import { useWalletConnection } from 'hooks/useWalletConnection'
-
-export const BlockExplorerLink: React.FC<Props> = (props) => {
-  const { networkId: networkIdFixed } = props
-  const { networkIdOrDefault: networkIdWallet } = useWalletConnection()
-
-  const networkId = networkIdFixed || networkIdWallet
-
-  return <BlockExplorerLink {...props} networkId={networkId} />
-}
+export const BlockExplorerLink = BlockExplorerLinkWalletConnection
